phoebe-lab-ui/requirements.txt
fastapi
uvicorn[standard]
gunicorn
requests
psutil
phoebe
pyzmq
nicegui
plotly
